Replace ClassNames, making it safer and faster
This stops using hashes for equality between classes.
Moves to using an incrementing id. This loses deterministic ids for each
class, which is unfortunate, but it was a poor idea to use hashes
anyway.
This also moves to using the bytes directly rather than converting to
strings, which was very much unneeded in the vast majority of
situations. Typically, you only need to do such when displaying or
accessing files. Now, we finally only do it in those scenarios.