Split Instance into Static and Reference Values
This makes Instance have two levels.
The first is merely a static class.
The next is for Reference values, which is itself a separate
enumeration.
Splitting this into two variants with values having a separate variant
allows us to have greater type safety. Another nice part of this whole
thing is that it lets us represent that there can't be a static class
gc-reference on the runtime stack.
Another part is a more generic upcasting mechanism for Instance-related
types.
Minor other fixes/improvements were made.