Add basic JNI and native methods support
This is more just the backbone that is needed to make it work, as
calling any native method - at the moment - that does anything remotely
useful is likely to crash because of the nullptr functions.
This also only lets you call basic functions with no args and no return
value.