Enable usage of NativeLibraries in multiple threads
This wraps the usage of NativeLibraries inside an RwLock and an Arc so
that it can be accessed from multiple threads in a reasonable manner.
RwLock was chosen because we do not often load new libraries (a mutable
action), but we do often lookup symbols (immutable).
This commit also switches over a static version where symbols are
assumed to live forever, which avoids complicated lifetimes, and was
deemed fine since we will most-likely rarely need to unload.