Use wrapper Env structure; Fix JNI ptr
We switch to using an Env structure which wraps a lot of the common
classes that are used and passed around.
This helps minimize typing of arguments.
Though, it is not really useful for having helper functions on it since
we don't have partial borrows and so it would borrow whole structure.
The main benefit it brings is that it lets us avoid weird complex
structures where ThreadData owns the JNI (as was the initial idea) but
would need to be given all that data (including ThreadData, which has to
include itself for further native method calls....).

Removed old JNIEnv, using Env as the replacement.
Rename JNINativeInterface -> NativeInterface
Java Native Interface Native Interface :)