Add StaticForm reference for Class<T> types
This has so it is a special reference to avoid weird special handling
for normal ClassInstances.
We also have it separate from the StaticClassInstance so that it avoids
having their static fields on it, which would also complicate typical
handling. This does mean one extra alloc for things which use Class<T>,
unfortunately.