Fix field initialization for tree of classes
Before we were simply initializing the fields of the bottommost class,
which is obviously incorrect. We also need to initialized all the super
class' fields.
Another issue we had was that you can have a class with some field 'a'
that extends another class with a field name 'a' (same type or not) and
they'll count as different fields.
Due to that, we use the class id as part of the key. We also had to
switch to using IndexMap so that we can use the lovely Equivalence
trait.