Fix fieldids being incorrect by including ClassId
When switching over from fieldnames -> ?fieldids? -> fieldindexes, we lost
the idea that fieldindexes (like names!) are per class and so have to be
differentiated between one that looks the same in the parent class.
Thus, this commit simply adds a classid field to keep track of, in the
style of MethodId