Add own definition of String class
This adds an implementation of the String class for the RhoJVM.
The primary reason for this is that we need access to the internal data
of the string, and making a cross-jvm-lib way of doing this was
basically impossible.
This is unfortunate (even if it is understandable why the JVM doesn't
provide any standard way of accessing the internals of a String) because
our implementation may not be as efficient and currently has issues of
a few missing function implementations and not supporting locales
properly.