Add support for reflection with fields
This implements java/lang/Field, in terms of an internal
rho/InternalField java type.
Implements java/lang/Class#getDeclaredField, though the implementation of that is
rather bad but can be improved as we add exception support properly to
native functions.
As well, this adds the proper support for Rust callbacks (in extern "C"
style) being used as overrides, of which class_get_declared_field is one
of them.