Improve formatter
This is honestly a minor change compared to the amount of work it took
to decide to do it.
I've had the issue for a while that tracing is both overkill, underkill, and a poor fit for this project.
- Its default formatting is overly verbose for anything that goes a few
  scopes deep, which running functions obviously does
- It provides json message output, but I have no clue of any tools to
  use them, and I found it challenging to find one. Especially one that
would be embedded into the editor (VSCode), which would minimize the
pain of actually reading the logs
- Filtering is rather basic. Env filter seems to be more about isolating
  specific parts of interest, while what I want is more filter out
specifics that are probably useless (like calling String::<init>)
- I can't use with_filter on my logging??? I still don't know why this
  is so, but it panics with an error message about it not being
supported.
I tried slog, but passing around a log is a pain, and making the api
work well when I seemingly have to swap out the log instance is hard.
I first tried simply putting it on my main structure, Env. The issue,
however, is that in eval_method we'd want to go into a layer. So we'd
swap out the log on Env with a .new'd one.
However, any time there is an error (for a result), it would return and
there isn't a great way in Rust to restore the original
indentation-level/log/scope.