Improve native method handling
This adds a macro to make it easier to generate code for different
native method calling, though it is not quite finished as it does not
make generating the permutations easy.
This also fixes using the JValue enum where I really shouldn't, due to a
confusion about whether it was being used as the general purpose return
type.
The native method calling is rather annoying in Rust, and I do wonder
how they typically implement it in other jvms.