Fix SuperClassFileIterator's behavior with arrays
The issue was that it was assuming that if it loaded a class file, it
would thus have one, which was a valid assumption early in the codebase.
However, the current behavior with arrays is to silently do nothing,
which made it error when it was inevitably passed an array as part of
casting checks.