Have get_field_dest look in super classes
This was an oversight in the initial implementation, because many
accesses to fields have the actual class that it is on as the class in
the field info. However, we have to look in the super classes if it does
not exist.
Unfortunately, this makes so we have to allocate the field name, or have
to reget the field repeatedly.