Add support for Unsafe allocating
This keeps track of the allocated blocks off of the GC, because there
isn't really any meaningful sense that we garbage collect them without
hiding data with the returned longs (since we're a vm, we could do a
special structure that is completely opaque to the java code), but that
would be a lot of extra instrumentation for likely little benefit.
This is pretty unsafe, but should hopefully be sound since it is not too
complex and through the apis we only allow modifying when we're a
mutable reference anyway.