Correct InvokeVirtual behavior
It appears like the class given doesn't actually signify where the 'root' definition of the target method is, since it resolves the method which looks at super classes and super interfaces. This means that this is maybe still too strict.
It also would be easy to make it too lenient.