Add the ability to abort on UnsupportedOperationExceptions
This is mostly to make developing easier, as they are what is used to
say that we haven't yet gotten a function implemented. Typically,
constructing them fully and going through throwing can execute a lot of
code which can obscure where it occurred. This would be less of an issue
if we put the name of the function in the text, but we don't.