More accurate and robust against circularity Class<T>
Previously, we always initialized the class when creating the Class<T>
instance. That had some benefits, but it was invalid. For simple
programs, this was fine, but in more complex programs it could run into
circularity issues, because just getting a Class<T> does not mean it
initializes the class.
We ran into this issue when loading a class as it wanted to use the
static variables of another class, but those are only initialized after
some setup which we were still in the process of.