Directly construct Java String
This is somewhat faster since we no longer have to do all the evaluate
method setup.
It is also simpler to do.
The main downside is that it makes it harder for us to use another JVM's
string implementation sice it may not work like ours. However, this is
probably fine, and the way we had support was hacky in the first place.
This also helps avoid circularity issues since we're no longer calling
constructors which could use the empty string (which would then call a
constructor that uses the empty string). This is what the hack fixed,
but it is no longer needed.
Also empty string was moved to the interner rather than a field on
state.